// client/src/models/transactions.ts
  export interface StatementSummary {
    openingBalance: string;
    totalCredits: string;
    totalDebits: string;
    closingBalance: string;
    minBalance: string;
    minDayEndBalance: string;
    daysInNegative: number;
    searchPeriodStart: string;
    searchPeriodEnd: string;
    transactionsStartDate: string;
    transactionsEndDate: string;
    creditsCategorised: string;
    creditsUncategorised: string;
    debitsCategorised: string;
    debitsUncategorised: string;
    internalTransferDebits: string;
    internalTransferCredits: string;
    accountMovement: string;
    dayEndBalances: DayEndBalance[];
  }
  
  export interface DayEndBalance {
    date: string;
    balance: string;
  }